Game of NIM
Welcome to the game of nim.  This first version just plays the game and employs two human players.

This nim game version starts with 21 sticks and the player selects 1, 2 or 3 sticks. If a player does not select 1, 2 or 3 sticks an error occurs and they again are asked to select 1, 2 or 3 sticks.   This continue until all the sticks are gone,  The player selecting the last stick loses.  the game then asks if the players would like to play another game.